Fix branch coverage for yield statements. #308 #324
Turns out the "call" and "return" trace events are really "start frame" and
"end frame".  They happen not only when functions are entered and left, but
when generators yield and resume.  We aren't interested in arcs into and out
of yield statements, so the trace functions look more closely to see what's
really happening, and record an arc in human-friendly terms.

Thanks for Mickie Betz for pushing on this bug, although her code is no
longer here. :(