Move try_execfile.py to an unambiguous place
Where it was, we were counting on "tests" referring to coverage.py's own tests
directory, but a -e install of a helper like unittest-mixins could provide a
tests directory, which prevented the imports we needed for the tests.

Moving try_execfile.py to a new subdirectory with a less-common name solves
the problem.

--HG--
rename : tests/try_execfile.py => tests/modules/process_test/try_execfile.py