finally happens before return in a try #493
In a try block such as:

  if expr:
    try:
      return
    finally
      print
  pass

the print happens before the return and cannot be followed by
pass. The general case is that when the body/else/handlers in a try
block all have return, break etc., the code behind finally: has no arc
to the statement following the try block.

close #493

--HG--
branch : issue-493-2