Simplify stack management in CTracer
"file_data" used to be borrowed from data, but that was confusing. Now it's
owned.  We used to have a struct member which was a copy of the current stack
entry.  That just made it harder to reason about reference counting.  Now we
have a pointer to the entry on the stack.