Update version.py to follow semver rules
In its current state _make_version would generate version strings that were not compatiable with semver rules. This was causing issues in that pre-release versions of the code were being downloaded.

The code has been updated to always include minor version and to use a hyphen to separate pre-release versions.