Canonicalize the XML output
https://bugs.python.org/issue34160 added retaining the user's attribute
order to the XML output, which removed the sorting that used to happen.
This broke our XML tests, which compare against saved gold files. This
adds in a rough-and-ready canonicalization to avoid the problem.

Maybe the core devs will eventually support a sort_attributes option,
and I can get rid of this.