Return to the original directory after exec'ing a file
If the file that is exec'd chdirs to a directory that doesn't exist
at the end of the execution, then we will fail to connect to the
SQLite database (due to a failing `os.getcwd` command).

We can easily fix this if we ensure we are in a directory that
exists after executing the foreign code. Returning to the original
directory seems to be a sensible choice.