Exit arcs have to be corrected to first lines
We've long remapped line numbers to the first line of a multi-line
statement.  But exit line numbers (negative numbers) were not remapped.
This meant we were needlessly chasing weirdnesses in implementations.

But the actual results of running coverage always remapped results to
the first line, so there's no point in tracking the unmapped line
numbers in our tests.