CmdLineStdoutTest::test_cmd_help: test for at least 20 lines
The number of lines in the help output of a command depends on the terminal
size. The smaller the more line breaks.

The minimum number of lines for the current help message is 23. Currently we
are checking for at least 30 lines, yielding to failures on large terminals.

Reduce the number (currently 30) to 20 to have some leeway for the future.