test: adapt to pytest assertion messages
Bare "assert" statements don't produce the same assertion message as
self.assertEqual did: they don't include the two values compared.  For
some of our own asserts, add back the detailed message.  For some checks
of asserts, it's fine that the values are missing because the longer
messsage includes the information.