test: force VirtualenvTest onto one worker to save time
VirtualenvTest is slow because it has a session-scoped fixture that takes 10s to
run.  If all of those tests go to the same worker, we can reuse that fixture. If
they go to different workers, we have to spend that time more than once.

This is a hack which seems to work, but no guarantees into the future. Also, I
don't know why the VirtualenvTests aren't run first on their worker.

Time saved: about 10%, from ~50s to ~45s.