coverage run: add an output option (-o)
Similar to coverage report/xml/json, coverage run now has a -o option to
configure the file the SQLite coverage data will be stored in. The
default is still ".coverage", as previously.

Since data_file now needs to be specified in cmdline.py,
_DEFAULT_DATAFILE needs to be exposed (to avoid hacks like constructing
the argument dictionary manually). Rename it to non-undescore and create
an alias for it under the previous name (just in case someone used it
anyway).