speed up python tracer with frame.f_trace_lines = False
use the python >= 3.7 feature of being able to disable line tracing by
setting the frame attribute f_trace_lines to False. This can be used for
the frames of functions that we aren't collecting coverage information
for (eg those functions in the stdlib).

This speeds up the pure python tracer in CPython by ~9x and in PyPy by
80% on a coverage run of one realistic project that I tried.