build: fix the branch protection gate check @ GHA
This adds a GHA job that reliably determines if all the required
dependencies have succeeded or not.

It is now in use in aiohttp (and other aio-libs projects), CherryPy,
some of the Ansible repositories, all of the jaraco's projects (like
`setuptools`, `importlib_metadata`), some of hynek's projects, some
PyCQA, PyCA, PyPA and pytest projects, a few AWS Labs projects (to my
surprise).

The story behind this is explained in more detail at
https://github.com/marketplace/actions/alls-green#why.