Strip trailing slash for relative source paths
This fixes an issue introduced in 45cf7936ee605cfe06f7f5967a72a73198960120
where using `relative_files=True` and `src` with a trailing slash,
the source inserted as `<sources>` in the XML report would also have
a trailing slash. This also fixes an issue introduced in the same
commit where an empty `<source>` would be inserted as well for cases
where the `src` has a trailing slash.