Add regression test for path w/ trailing wildcard
This patch demonstrates a corner case in the path glob matcher.
Specifically, it documents how a single trailing asterisk is supposed
to be treated as opposed to a double asterisk.
With [[1]], a trailing `/*` is interpreted as an equivalent of `/**`.
The commit add a case that shows that `/*` shouldn't be greedy as
described in the docs [[2]][[3]].

See also the observations in the bug report ticket [[4]].

[1]: https://github.com/nedbat/coveragepy/commit/ec6205a8de972af6a09453235d02a7ebea6aea8e
[2]: https://coverage.rtfd.io/en/stable/source.html#file-patterns
[3]: https://coverage.rtfd.io/en/7.2.7/migrating.html#migrating-to-coverage-py-7-x
[4]: https://github.com/nedbat/coveragepy/issues/1407#issuecomment-1631085209