refactor: html report files start with z_
When I download an HTML report and unzip it, I need to open the
directory and find the index.html file.  With a d_ prefix, index.html is
almost the last file in the directory.  Change the prefix to z_ so that
now index.html is nearly the first. More convenient for me.