fix: lcov report improvements #1850 (#1851)
* refactor: don't need this aux variable anymore

* feature: add arc_description method for #1850

* [lcov] Refactor LcovReporter.lcov_file

Split the bulk of the code in LcovReporter.lcov_file out into two free helper
functions, lcov_lines and lcov_arcs.  This is easier to read and will make it
easier to do future planned changes in a type-safe manner.

No functional changes in this commit.

* [lcov] Improve reporting of branch destinations.

The branch field of a BRDA: record can be an arbitrary textual label.
Therefore, instead of emitting meaningless numbers, emit the string
“to line <N>” for ordinary branches (where <N> is the arc destination
line, and “to exit” for branches that exit the function.  When there is
more than one exit arc from a single line, provide the negated arc
destination as a disambiguator.

Thanks to Henry Cox (@henry2cox), one of the LCOV maintainers, for
clarifying the semantics of BRDA: records for us.

* Implement function coverage reporting in lcov reports.

Quite straightforward: a function has been executed if any of its region’s
lines have been executed.

* [lcov] Ignore vacuous function regions.

Should fix the test failures with pypy pretending to be python 3.8.

* Adjust test expectations for lcov reports generated under PyPy 3.8.

There is a bug somewhere, in which if we collect data in --branch mode under
PyPy 3.8, regions for top-level functions come out of the analysis engine with
empty lines arrays.  The previous commit prevented this from crashing the lcov
reporter; this commit adjusts the tests of the lcov reporter so that we expect
the function records affected by the bug to be missing.

I don’t think it’s worth trying to pin down the cause of the bug, since Python
3.8 is approaching end-of-life for both CPython and PyPy.

* Split up test_multiple_exit_branches into 3 tests exercising #1852.

* add a fourth case to the tests for #1852

* Revise lcovreport.lcov_arcs using the new arc_description API.

Doesn’t quite work yet, see discussion in #1850.

* fix: forgot to add arc_descriptions to Python filereporter

* refactor: make arc possibilities and arcs executed available

* fix: executed_branch_arcs should limit itself to parsed possible arcs

* Use tests.helpers.xfail_pypy38 for lcov tests that fail with pypy 3.8.

This replaces the custom fn_coverage_missing_in_pypy_38 logic that was
used in earlier commits.

* tests: split xfail_pypy38 decorator into _older_ and _all_ variants

The lcov output tests that are affected by bugs in PyPy 3.8, fail with the
current version of PyPy 3.8 (7.3.11), unlike the other tests annotated with
@xfail_pypy38.  Split this decorator into two variants, xfail_older_pypy38
(used for all the tests that were labeled xfail_py38 prior to this patchset)
and xfail_all_pypy38 (used for the lcov output tests).

---------

Co-authored-by: Ned Batchelder <ned@nedbatchelder.com>