test: don't try to make pth files when invoked from pth #2011
When we're invoked by a .pth file, site.py hasn't fully run, so we try
to make a .pth file in the system Python instead of the virtualenv.
System Pythons are often not writable.

Also try to create this kind of not-writable arrangement in CI to find
problems like this in the future.