Move the core of the combine logic to be entirely in SQL
This shaves roughly 40% of the runtime off of pyca/cryptography's combine.

A major limitation is that it currently doesn't work with in-memory sqlite databases, I'm not sure if there's a good way to do that.

(cherry picked from commit 270ed94a826f23fd7f6b1fae7da41b88a565e170)